[IMPROVE] change return type of price and qty from Float to Integer

+++ src/main/java/com/enigma/wmb_api/dto/request/BillDetailRequest.java
@@ -16,5 +16,5 @@ public class BillDetailRequest {
 
     @NotNull
     @Min(value = 1)
-    private Float qty;
+    private Integer qty;
 }
